Completed Challenge 2: Implement average response time.
Now, sending a GET request to endpoint /metrics/conversations/2
returns long conversationID and a double avgResponseMs. We also
calculated a new metric: inquiryToResponseRatio, that returns how
many inquiries an end user has vs how many responses a Team Member
provided. In the next commit, we will change the builder of the
ConversationResponseMetric so it returns inquiryToResponseRatio as well.